+++ src/assert_arrays.sh
@@ -2,8 +2,10 @@
 
 function assert_array_contains() {
   local expected="$1"
+  local test_fn
+  test_fn="$(helper::find_test_function_name)"
   local label
-  label="$(helper::normalize_test_function_name "${FUNCNAME[1]}")"
+  label="$(helper::normalize_test_function_name "$test_fn")"
   shift
 
   local actual=("${@}")
@@ -19,7 +21,10 @@ function assert_array_contains() {
 
 function assert_array_not_contains() {
   local expected="$1"
-  label="$(helper::normalize_test_function_name "${FUNCNAME[1]}")"
+  local test_fn
+  test_fn="$(helper::find_test_function_name)"
+  local label
+  label="$(helper::normalize_test_function_name "$test_fn")"
   shift
   local actual=("$@")